+++ service/src/main/kotlin/io/provenance/explorer/domain/core/Bech32.kt
@@ -1,19 +1,13 @@
 package io.provenance.explorer.domain.core
 
 import com.google.common.io.BaseEncoding
-
-infix fun Int.min(b: Int): Int = b.takeIf { this > b } ?: this
-@ExperimentalUnsignedTypes
-infix fun UByte.shl(bitCount: Int) = ((this.toInt() shl bitCount) and 0xff).toUByte()
-@ExperimentalUnsignedTypes
-infix fun UByte.shr(bitCount: Int) = (this.toInt() shr bitCount).toUByte()
+import java.io.ByteArrayOutputStream
 
 /**
  * Given an array of bytes, associate an HRP and return a Bech32Data instance.
  */
-@ExperimentalUnsignedTypes
-fun ByteArray.toBech32Data(hrp: String = Bech32.PROVENANCE_MAINNET_ACCOUNT_PREFIX) =
-    Bech32Data(hrp, Bech32.convertBits(this, 8, 5, true))
+
+fun ByteArray.toBech32Data(hrp: String = Bech32.PROVENANCE_MAINNET_ACCOUNT_PREFIX) = Bech32Data(hrp, this)
 
 /**
  * Using a string in bech32 encoded address format, parses out and returns a Bech32Data instance
@@ -23,36 +17,24 @@ fun String.toBech32Data() = Bech32.decode(this)
 /**
  * Bech32 Data encoding instance containing data for encoding as well as a human readable prefix
  */
-data class Bech32Data(val hrp: String, val fiveBitData: ByteArray) {
-
-    /**
-     * The encapsulated data as typical 8bit bytes.
-     */
-    @ExperimentalUnsignedTypes
-    val data = Bech32.convertBits(fiveBitData, 5, 8, false)
+data class Bech32Data(val hrp: String, val data: ByteArray) {
 
     /**
      * The encapsulated data returned as a Hexadecimal string
      */
-    @ExperimentalUnsignedTypes
+
     val hexData = BaseEncoding.base16().encode(this.data)
 
     /**
      * Address is the Bech32 encoded value of the data prefixed with the human readable portion and
      * protected by an appended checksum.
      */
-    val address = Bech32.encode(hrp, fiveBitData)
-
-    /**
-     * Checksum for encapsulated data + hrp
-     */
-    val checksum = Bech32.checksum(this.hrp, this.fiveBitData.toTypedArray())
+    val address = Bech32.encode(hrp, data)
 
     /**
      * The Bech32 Address toString prints state information for debugging purposes.
      * @see address() for the bech32 encoded address string output.
      */
-    @ExperimentalUnsignedTypes
     override fun toString(): String {
         return "bech32 : ${this.address}\nhuman: ${this.hrp} \nbytes: ${this.hexData}"
         /*
@@ -60,6 +42,22 @@ data class Bech32Data(val hrp: String, val fiveBitData: ByteArray) {
         human: provenance
         bytes: 41A87604E1192F1959869D982DB5F706C0099006
          */
+    }
+
+    /** equals implementation for a Bech32Data object. */
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+        other as Bech32Data
+        return this.hrp == other.hrp &&
+            this.data.contentEquals(other.data)
+    }
+
+    /** equals implementation for a Bech32Data object. */
+    override fun hashCode(): Int {
+        var result = hrp.hashCode()
+        result = 31 * result + this.data.contentHashCode()
+        return result
     }
 }
 
@@ -78,22 +76,27 @@ class Bech32 {
         const val charset = "qpzry9x8gf2tvdw0s3jn54khce6mua7l"
         val gen = intArrayOf(0x3b6a57b2, 0x26508e6d, 0x1ea119fa, 0x3d4233dd, 0x2a1462b3)
 
-        const val PROVENANCE_MAINNET_PREFIX                             = "pb"
+        // Mainnet account prefixes
+        const val PROVENANCE_MAINNET_PREFIX                   = "pb"
         const val PROVENANCE_MAINNET_ACCOUNT_PREFIX           = PROVENANCE_MAINNET_PREFIX
         const val PROVENANCE_MAINNET_VALIDATOR_ACCOUNT_PREFIX = PROVENANCE_MAINNET_PREFIX + "valoper"
         const val PROVENANCE_MAINNET_CONSENSUS_ACCOUNT_PREFIX = PROVENANCE_MAINNET_PREFIX + "valcons"
 
         // Test net account prefixes are broken out separately so keys/accounts used for test can be easily identified
-
-        const val PROVENANCE_TESTNET_PREFIX                             = "tp"
+        const val PROVENANCE_TESTNET_PREFIX                   = "tp"
         const val PROVENANCE_TESTNET_ACCOUNT_PREFIX           = PROVENANCE_TESTNET_PREFIX
         const val PROVENANCE_TESTNET_VALIDATOR_ACCOUNT_PREFIX = PROVENANCE_TESTNET_PREFIX + "valoper"
         const val PROVENANCE_TESTNET_CONSENSUS_ACCOUNT_PREFIX = PROVENANCE_TESTNET_PREFIX + "valcons"
 
+        // NFT (Scope) prefixes
+        const val SCOPE_PREFIX = "scope"
+        const val SESSION_PREFIX = "session"
+        const val RECORD_PREFIX = "record"
+        const val SCOPE_SPEC_PREFIX = "scopespec"
+        const val CONTRACT_SPEC_PREFIX = "contractspec"
+        const val RECORD_SPEC_PREFIX = "recspec"
 
-        /**
-         * Decodes a Bech32 String
-         */
+        /** Decodes a Bech32 String */
         fun decode(bech32: String): Bech32Data {
             require(bech32.length in MIN_VALID_LENGTH..MAX_VALID_LENGTH) { "invalid bech32 string length" }
             require(bech32.toCharArray().none { c -> c.toInt() < MIN_VALID_CODEPOINT || c.toInt() > MAX_VALID_CODEPOINT })
@@ -115,85 +118,63 @@ class Bech32 {
             val actualSum = checksum(hrp, dataBytes.dropLast(CHECKSUM_SIZE).toTypedArray())
             require(1 == polymod(expandHrp(hrp).plus(dataBytes.map { d -> d.toInt() }))) { "checksum failed: $checkBytes != $actualSum" }
 
-            return Bech32Data(hrp, dataBytes.dropLast(CHECKSUM_SIZE).toByteArray())
+            return Bech32Data(hrp, convertBits(dataBytes.dropLast(CHECKSUM_SIZE).toByteArray(), 5, 8, false))
+        }
+
+        /**
+         * Encodes the provided hrp and data to a Bech32 address string.
+         * @param hrp the human readable portion (prefix) to use.
+         * @param eightBitData an array of 8-bit encoded bytes.
+         */
+        fun encode(hrp: String, eightBitData: ByteArray) =
+            encodeFiveBitData(hrp, convertBits(eightBitData, 8, 5, true))
+
+        /** Encodes 5-bit bytes (fiveBitData) with a given human readable portion (hrp) into a bech32 string. */
+        private fun encodeFiveBitData(hrp: String, fiveBitData: ByteArray): String {
+            return (fiveBitData.plus(checksum(hrp, fiveBitData.toTypedArray()))
+                .map { b -> charset[b.toInt()] }).joinToString("", hrp + "1")
         }
 
         /**
          * ConvertBits regroups bytes with toBits set based on reading groups of bits as a continuous stream group by fromBits.
          * This process is used to convert from base64 (from 8) to base32 (to 5) or the inverse.
          */
-        @ExperimentalUnsignedTypes
-        fun convertBits(data: ByteArray, fromBits: Int, toBits: Int, pad: Boolean): ByteArray {
+        private fun convertBits(data: ByteArray, fromBits: Int, toBits: Int, pad: Boolean): ByteArray {
             require (fromBits in 1..8 && toBits in 1..8) { "only bit groups between 1 and 8 are supported"}
 
-            // resulting bytes with each containing the toBits bits from the input set.
-            val regrouped = arrayListOf<Byte>()
+            var acc = 0
+            var bits = 0
+            val out = ByteArrayOutputStream(64)
+            val maxv = (1 shl toBits) - 1
+            val maxAcc = (1 shl (fromBits + toBits - 1)) - 1
 
-            var nextByte = 0.toUByte()
-            var filledBits = 0
-
-            data.forEach { d ->
-                // discard unused bits.
-                var b = (d.toUByte() shl (8 - fromBits))
-
-                // How many bits remain to extract from input data.
-                var remainFromBits = fromBits
-
-                while (remainFromBits > 0) {
-                    // How many bits remain to be copied in
-                    val remainToBits = toBits - filledBits
-
-                    // we extract the remaining bits unless that is more than we need.
-                    val toExtract = remainFromBits.takeUnless { remainToBits < remainFromBits } ?: remainToBits
-                    check(toExtract >= 0) { "extract should be positive"}
-
-                    // move existing bits to the left to make room for bits toExtract, copy in bits to extract
-                    nextByte = (nextByte shl toExtract) or (b shr (8 - toExtract))
-
-                    // discard extracted bits and update position counters
-                    b = b shl toExtract
-                    remainFromBits -= toExtract
-                    filledBits += toExtract
-
-                    // if we have a complete group then reset.
-                    if (filledBits == toBits) {
-                        regrouped.add(nextByte.toByte())
-                        filledBits = 0
-                        nextByte = 0.toUByte()
-                    }
+            for (b in data) {
+                val value = b.toInt() and 0xff
+                if ((value ushr fromBits) != 0) {
+                    throw IllegalArgumentException(String.format("Input value '%X' exceeds '%d' bit size", value, fromBits))
+                }
+                acc = ((acc shl fromBits) or value) and maxAcc
+                bits += fromBits
+                while (bits >= toBits) {
+                    bits -= toBits
+                    out.write((acc ushr bits) and maxv)
                 }
             }
-
-            // pad any unfinished groups as required
-            if (pad && filledBits > 0) {
-                nextByte = nextByte shl (toBits - filledBits)
-                regrouped.add(nextByte.toByte())
-                filledBits = 0
-                nextByte = 0.toUByte()
+            if (pad) {
+                if (bits > 0) {
+                    out.write((acc shl (toBits - bits)) and maxv)
+                }
+            } else if (bits >= fromBits || ((acc shl (toBits - bits)) and maxv) != 0) {
+                throw IllegalArgumentException("Could not convert bits, invalid padding")
             }
-
-            // check for any incomplete groups that are more than 4 bits or not all zeros
-            require (filledBits == 0 || (filledBits <= 4 && nextByte == 0.toUByte())) { "invalid incomplete group" }
-
-            return regrouped.toByteArray()
+            return out.toByteArray()
         }
 
-        /**
-         * Encodes data 5-bit bytes (data) with a given human readable portion (hrp) into a bech32 string.
-         * @see convertBits for conversion or ideally use the Bech32Data extension functions
-         */
-        fun encode(hrp: String, fiveBitData: ByteArray): String {
-            return (fiveBitData.plus(checksum(hrp, fiveBitData.toTypedArray()))
-                .map { b -> charset[b.toInt()] }).joinToString("", hrp + "1")
-        }
-
-        /**
-         * Calculates a bech32 checksum based on BIP 173 specification
-         */
-        fun checksum(hrp: String, data: Array<Byte>): ByteArray {
+        /** Calculates a bech32 checksum based on BIP 173 specification */
+        private fun checksum(hrp: String, data: Array<Byte>): ByteArray {
             val values = expandHrp(hrp)
                 .plus(data.map { d -> d.toInt() })
-                .plus(Array<Int>(6){ _ -> 0}.toIntArray())
+                .plus(Array(6){ 0 }.toIntArray())
 
             val poly = polymod(values) xor 1
 
@@ -202,23 +183,19 @@ class Bech32 {
             }.toByteArray()
         }
 
-        /**
-         * Expands the human readable prefix per BIP173 for Checksum encoding
-         */
-        fun expandHrp(hrp: String) =
+        /** Expands the human readable prefix per BIP173 for Checksum encoding */
+        private fun expandHrp(hrp: String) =
             hrp.map { c -> c.toInt() shr 5 }
                 .plus(0)
                 .plus(hrp.map { c -> c.toInt() and 31 })
                 .toIntArray()
 
-        /**
-         * Polynomial division function for checksum calculation.  For details see BIP173
-         */
-        fun polymod(values: IntArray): Int {
+        /** Polynomial division function for checksum calculation.  For details see BIP173 */
+        private fun polymod(values: IntArray): Int {
             var chk = 1
-            return values.map {
+            return values.map { v ->
                 val b = chk shr 25
-                chk = ((chk and 0x1ffffff) shl 5) xor it
+                chk = ((chk and 0x1ffffff) shl 5) xor v
                 (0..4).map {
                     if (((b shr it) and 1) == 1) {
                         chk = chk xor gen[it]