+++ service/src/main/kotlin/io/provenance/explorer/web/v2/NftController.kt
@@ -1,0 +1,59 @@
+package io.provenance.explorer.web.v2
+
+import io.provenance.explorer.service.NftService
+import io.provenance.explorer.web.BaseController
+import io.swagger.annotations.Api
+import io.swagger.annotations.ApiOperation
+import org.springframework.http.MediaType
+import org.springframework.http.ResponseEntity
+import org.springframework.validation.annotation.Validated
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.bind.annotation.RestController
+import javax.validation.constraints.Min
+
+@Validated
+@RestController
+@RequestMapping(path = ["/api/v2/nft"], produces = [MediaType.APPLICATION_JSON_VALUE])
+@Api(value = "NFT controller", produces = "application/json", consumes = "application/json", tags = ["NFTs"])
+class NftController(private val nftService: NftService) : BaseController() {
+
+    @ApiOperation("Returns NFT detail for address")
+    @GetMapping("/{uuid}")
+    fun getNftDetail(@PathVariable uuid: String) = ResponseEntity.ok(nftService.getScopeByUuid(uuid))
+
+    @ApiOperation("Returns paginated list of NFTs")
+    @GetMapping("/all")
+    fun getAllNfts(
+        @RequestParam(required = false, defaultValue = "1") @Min(1) page: Int,
+        @RequestParam(required = false, defaultValue = "10") @Min(1) count: Int
+    ) =
+        ResponseEntity.ok(nftService.getAllScopes(page, count))
+
+    @ApiOperation("Returns NFTs for owning address")
+    @GetMapping("/owner/{address}")
+    fun getNftsByOwningAddress(
+        @PathVariable address: String,
+        @RequestParam(required = false, defaultValue = "1") @Min(1) page: Int,
+        @RequestParam(required = false, defaultValue = "10") @Min(1) count: Int
+    ) = ResponseEntity.ok(nftService.getScopesForOwningAddress(address, page, count))
+
+    @ApiOperation("Returns MetadataAddress obj for bech32 addr")
+    @GetMapping("/address/{addr}")
+    fun getMetadataAddress(@PathVariable addr: String) = ResponseEntity.ok(nftService.translateAddress(addr))
+
+
+
+}
+
+
+// Per NFT (NEW PAGE)
+//- the NFT
+//- whatever is under the NFA (I dont know yet)
+//- NFT transactions
+//  - again no clue what this would look like
+//
+//Non-Fungible Token - NFT (NEW PAGE)
+//- list of NFTs